+++ bootique-agrest-jakarta-junit5/src/main/java/io/bootique/agrest/junit5/AgTester.java
@@ -29,8 +29,14 @@ public class AgTester {
         };
     }
 
+    /**
+     * Adds a requesting customization function to the tester.
+     */
     public AgTester customizeRequest(Consumer<Invocation.Builder> requestCustomizer) {
-        this.requestCustomizer = Objects.requireNonNull(requestCustomizer);
+        Objects.requireNonNull(requestCustomizer);
+        this.requestCustomizer = this.requestCustomizer != null
+                ? this.requestCustomizer.andThen(requestCustomizer)
+                : requestCustomizer;
         return this;
     }